reorganized project for future scalability
