--- src/Info.jsx
@@ -1,13 +1,0 @@
-
-import "./Info.css"
-
-export default function Info() {
-    return (
-        <div className="body-text">
-            <div className="body-text-card">
-                <h1>Bienvenido!</h1>
-                <p>Sitio web creado para generar exámenes y exportarlos</p>
-            </div>
-        </div>
-    )
-}
