+++ src/pages/ExamGenerator.jsx
@@ -1,0 +1,17 @@
+import Header from "../components/Header"
+import Info from "../components/Info"
+import Body from "../components/Body"
+// import Footer from "../components/Footer";
+
+export default function ExamGenerator() {
+
+    return (
+        <>
+            <Header />
+            <Info />
+            <Body />
+            {/* <Footer /> */}
+        </>
+
+    )
+}
