+++ src/pages/VisualizeData.jsx
@@ -1,0 +1,8 @@
+export default function VisualizeData() {
+
+    return (
+        <>
+            <h1>Page for see data</h1>
+        </>
+    )
+}
