+++ src/resources/typeQuestions/QuestionSequence.jsx
