+++ src/resources/typeQuestions/styles/QuestionStyles.css
